Modular Version 1.0
**Major Change**
This branch/version splits the original one file Jarvis application, into a more organized project with different files for each component. This allows users to more easily access the options they need to change to make the application work after download, by editing the config.ini file. The application is now ran from the "main.py", instead of the "Jarvis Mark II.py".

This version also includes a Readme, Changelog, and a work in progress setup.bat to streamline the setup process.
