+++ config.ini
@@ -1,0 +1,18 @@
+[General]
+base_dir = C:\Users\ghost\Desktop\Coding\Python\Jarvis-Mark-II
+api_key = AIzaSyCMbXEdMHqh5WA0aH95qdAdpqVVTjURKx4
+voice_key = HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Speech\Voices\Tokens\MSTTS_V110_enGB_GeorgeM
+
+[Vosk]
+model_path = vosk-model-small-en-us-0.15
+
+[OBS]
+host = localhost
+port = 4455
+password = Tristen04
+
+[Sounds]
+sounds_dir = sounds
+
+[Music]
+music_dir = music
